requirements.txt
django
djangorestframework
selenium
pyautogui
webdriver-manager